subset an expression dataset by extracting genes based on their level of variation

subset an expression dataset by extracting genes based on their level of variation

VERBOSE(v, ...)

Arguments

dat
an expressionSet object

score
type of variation measure to use (can be 'mad', 'sd', or 'cv')

dir
direction of the gene ranking ('top': highest varying genes; 'bottom': lowest varying genes)

transform
transform the data before measuring variation

ngenes
number of top (or bottom) genes to extract

do.plot
plot the center vs. scale plot with selected genes highlighted

do.log
indicates whether to log the "x", "y", or "xy" axes, or "" none

Value

the filtered expressionSet object

Examples

# select the top 100 genes by median absolute deviation (mad) and also generate plot

# Use example data, for data set information: ?eSet.brca.100
data(eSet.brca.100)

#> Warning: data set ‘eSet.brca.100’ not found

variationFilter(eSet.brca.100,score='mad',ngenes=50,do.plot=TRUE)

#> Error in nrow(dat): object 'eSet.brca.100' not found